requirements.txt
fastapi
uvicorn
pdfminer.six
ebooklib
python-multipart
ruff
black
